requirements.txt
colorama==0.4.6
langchain==0.3.0
langchain_anthropic==0.2.0
langchain_core==0.3.0
langchain_google_genai==2.0.0
langchain_groq==0.2.0
langchain_mistralai==0.2.0
langchain_openai==0.2.0
python-dotenv==1.0.1
